% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Loads data from export}
\usage{
get_data()
}
\value{
list of dataframes from secuTrial export
}
\description{
Loads data from export
}
\seealso{
\code{\link{app_ui}}, \code{\link{app_srv}}
}
