% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_consistency.R
\name{mod_consistency_ui}
\alias{mod_consistency_ui}
\alias{mod_consistency_server}
\title{Consistency}
\usage{
mod_consistency_ui(id, label)

mod_consistency_server(input, output, session, data)
}
\arguments{
\item{id}{standard shiny id argument}

\item{label}{standard shiny label argument}

\item{input}{standard shiny input argument}

\item{output}{standard shiny output argument}

\item{session}{standard shiny session argument}

\item{data}{data for use in calculations}
}
\description{
In development
}
