% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_participant.R
\name{mod_participant_ui}
\alias{mod_participant_ui}
\alias{mod_participant_server}
\title{Participant information}
\usage{
mod_participant_ui(id, label)

mod_participant_server(input, output, session, data)
}
\arguments{
\item{id}{standard shiny id argument}

\item{label}{standard shiny label argument}

\item{input}{standard shiny input argument}

\item{output}{standard shiny output argument}

\item{session}{standard shiny session argument}

\item{data}{data for use in calculations}
}
\description{
In development
}
