% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recruitment2.R
\name{mod_recruitment2_ui}
\alias{mod_recruitment2_ui}
\alias{mod_recruitment2_server}
\title{Recruitment module Functions}
\usage{
mod_recruitment2_ui(id, label)

mod_recruitment2_server(
  input,
  output,
  session,
  data.randomized,
  locations,
  all_data
)
}
\arguments{
\item{id, input, output, session, label}{Internal parameters for {shiny}.}

\item{data.randomized}{reactive data containing randomization info
(see {mod_recruitment_prediction} for example without reactivity)}
}
\description{
This module presents a recruitment plot and table using the accrualPlot package
}
