% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recruitment_map.R
\name{mod_recruitment_map_ui}
\alias{mod_recruitment_map_ui}
\alias{mod_recruitment_map_server}
\title{Recruitment UI Functions including a map of the sites}
\usage{
mod_recruitment_map_ui(id, label)

mod_recruitment_map_server(input, output, session, dat, locations)
}
\arguments{
\item{id, input, output, session, label}{Internal parameters for {shiny}.}

\item{input, output, session}{Internal parameters for {shiny}.}

\item{dat}{reactive recruitment data}

\item{locations}{dataframe containing the locations of the sites.}
}
\description{
This module provides a leaflet map with markers for the sites
  together we a recruitment plot. Clicking on the map markers makes the relevant
  line on the recruitment plot become thicker. Hovering over the marker also
  shows various information on the site.
}
