% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_recruitment_prediction.R
\name{mod_recruitment_prediction_ui}
\alias{mod_recruitment_prediction_ui}
\alias{mod_recruitment_prediction_server}
\title{Recruitment prediction module}
\usage{
mod_recruitment_prediction_ui(id, label)

mod_recruitment_prediction_server(input, output, session, data)
}
\arguments{
\item{id}{standard shiny id argument}

\item{label}{standard shiny label argument}

\item{input}{standard shiny input argument}

\item{output}{standard shiny output argument}

\item{session}{standard shiny session argument}

\item{data}{data for use in calculations}
}
\description{
Predicts when a trial will reach it's target recruitment, based on recruitment to date
}
