% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_retention.R
\name{mod_retention_ui}
\alias{mod_retention_ui}
\alias{mod_retention_server}
\title{Retention}
\usage{
mod_retention_ui(id, label)

mod_retention_server(input, output, session, data)
}
\arguments{
\item{id}{standard shiny id argument}

\item{label}{standard shiny label argument}

\item{input}{standard shiny input argument}

\item{output}{standard shiny output argument}

\item{session}{standard shiny session argument}

\item{data}{data for use in calculations}
}
\description{
In development
}
