% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_skeleton.R
\name{mod_skeleton_ui}
\alias{mod_skeleton_ui}
\title{UI Function}
\usage{
mod_skeleton_ui(id, label)
}
\arguments{
\item{id, input, output, session, label}{Internal parameters for {shiny}.}
}
\description{
UI Function
}
