% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_visits.R
\name{mod_visits_ui}
\alias{mod_visits_ui}
\alias{mod_visits_server}
\title{Visits}
\usage{
mod_visits_ui(id, label)

mod_visits_server(input, output, session, data)
}
\arguments{
\item{id}{standard shiny id argument}

\item{label}{standard shiny label argument}

\item{input}{standard shiny input argument}

\item{output}{standard shiny output argument}

\item{session}{standard shiny session argument}

\item{data}{data for use in calculations}
}
\description{
In development
}
