% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruit}
\alias{plot_recruit}
\title{Plot recruitment curve}
\usage{
plot_recruit(data, date.var, centre.var)
}
\arguments{
\item{data}{dataframe}

\item{date.var}{variable containing dates}

\item{centre.var}{variable containing centers}
}
\value{
list of dataframes from secuTrial export
}
\description{
This function plots a recruitment curve by centers
}
