% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plot recruitment curve}
\usage{
plot_recruitment(data, date.var, centre.var)
}
\value{
list of dataframes from secuTrial export
}
\description{
This function plots a recruitment curve by centers
}
